+++ mobile_wallet/mobile_wallet.h
@@ -266,7 +266,7 @@ char *parameter_to_json(const char *input_ptr, uint8_t *success);
  * The input pointer must point to a null-terminated buffer, otherwise this
  * function will fail in unspecified ways.
  */
-char *create_pre_account_transaction(const char *input_ptr, uint8_t *success);
+char *create_account_transaction(const char *input_ptr, uint8_t *success);
 
 /**
  * Take a pointer to a NUL-terminated UTF8-string and return a NUL-terminated