+++ mobile_wallet/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Removed the function for signing any account transaction.
 - Added a function for converting serialized parameters for smart contract updates (as bytes) into JSON.
 - Added a function `create_account_transaction` for creating and signing an account transaction.
+- Added support for V3 schema
 
 ## 0.17.0
 - Added a function for signing any account transaction.